requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic>=2.7.0
numpy==1.23.5
opencv-python==4.5.5.64
opencv-contrib-python==4.5.5.64
einops==0.4.1
torch>=2.0.0
torchvision>=0.15.0
einops==0.4.1
pytorch-lightning==1.9.0
pillow==10.1.0
python-jose==3.3.0
matplotlib==3.5.1
imageio==2.25.0
scipy==1.10.1
omegaconf==2.3.0
transformers==4.19.2
pydantic-settings==2.8.1
open_clip-torch==2.23.0
xformers>=0.0.22